Add files via upload
Scenario: Imagine you are working on a 2D graphics processing tool that deals 
with image manipulation. Your task is to rotate an NxN image (represented by a 
matrix) 90 degrees clockwise in-place, which means you should not use extra 
memory. 
Task: Write a function that rotates the given matrix 90 degrees clockwise in
place.