Add files via upload
Welcome to the string game... the operations consist 1s and 0s... if it is 0 then the given word is multiplied by the same word its size times. for ex word is "a", if operations has 0 as first index value then word becomes aa, again 0 then aaaa. Suppose we have 1 then "a" output will be ab next character and so on ..